Rename files and using correct imports
