+++ components/index.tsx
@@ -5,5 +5,15 @@ import NavBar from "./Layout/NavBar/NavBar";
 import Footer from "./Layout/Footer/Footer";
 import MetaTags from "./Layout/MetaTags";
 import Banner from "./Banner/Banner";
+import OurCurrentEvents from "./Events/OurCurrentEvents";
 
-export { ContentContainer, Layout, ConstructionMessage, NavBar, Footer, MetaTags, Banner };
+export {
+  ContentContainer,
+  Layout,
+  ConstructionMessage,
+  NavBar,
+  Footer,
+  MetaTags,
+  Banner,
+  OurCurrentEvents,
+};
